[IDE] Dynamic Content System (pt1)
Just a lot. Tracks all modifications made by the IDE and allows users to
save or undo them. Modifies game content system to support this. Unify
all windows to base off of one central window.

+++ TSOClient/tso.files/formats/iff/chunks/OBJD.cs
@@ -431,7 +431,7 @@ namespace FSO.Files.Formats.IFF.Chunks
                 io.WriteUInt32(142);
                 foreach (var prop in VERSION_142_Fields)
                 {
-                    io.WriteUInt16(GetPropertyByName<ushort>(prop));
+                    io.WriteUInt16((ushort)GetPropertyByName<int>(prop));
                 }
                 for (int i = VERSION_142_Fields.Length; i < 105; i++)
                 {
